Level 4 Audio Visualiser attempt
The result failed. And the code was based on a non-final version of the group assignment, originally created in the p5 editor: https://editor.p5js.org/yslllll122/sketches/f51P-uJrt

According to the reference code, I tried to introduce the Level 4 animation, but there were some poor effects or no effects.